Change integer variables to double in Kalkulatormeh

+++ Kalkulatormeh.java
@@ -1,8 +1,8 @@
 public class Kalkulatormeh {
   public static void main(String [] args){
-    int angka1,angka2;
-    angka1 = 20;
-    angka2 = 30;
+    double angka1,angka2;
+    angka1 = 20d;
+    angka2 = 30d;
     System.out.println("(+) hasil tambah:" + (angka1 + angka2));
     System.out.println("(-) hasil kurang:" + (angka1 - angka2));
     System.out.println("(*) hasil kali:" + (angka1 * angka2));
